RelCAT: Initial Llama integration.
